requirements.txt
numpy
torch
torchvision
gymnasium[atari]
transformers
pyyaml
tensorboard
opencv-python
